refactor: delete a subdirectory of jwt
