+++ src/main/java/bitnagil/bitnagil_backend/auth/jwt/RefreshTokenRedisRepository.java
@@ -1,10 +1,8 @@
-package bitnagil.bitnagil_backend.auth.jwt.repository;
+package bitnagil.bitnagil_backend.auth.jwt;
 
 import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
-
-import bitnagil.bitnagil_backend.auth.jwt.entity.RefreshToken;
 
 public interface RefreshTokenRedisRepository extends CrudRepository<RefreshToken, String> {
     Optional<RefreshToken> findByRefreshToken(String refreshToken);
